Рефакторинг

+++ sources/Device/src/Settings/SettingsTransformation.cpp
@@ -43,7 +43,7 @@ void Settings::TransformForSave(uint8 d[1024])
     
     SaveEnum16(DisplayShowStringNavigation, (uint8)set.disp_showStringNavigation);
     SaveEnum16(DisplayENumMinMax, (uint8)set.disp_ENumMinMax);
-    SaveEnum16(DisplayModeAveraging, set.disp_modeAveraging);
+    SaveEnum16(DisplayModeAveraging, (uint8)set.disp_modeAveraging);
     SaveEnum16(DisplayENumAverage, set.disp_ENumAverage);
     SaveEnum16(DisplayENumAccum, set.disp_ENumAccum);
 }
